+++ src/commands/useful/word.js
@@ -1,0 +1,24 @@
+import randomWord from 'random-word';
+
+function word(client, evt) {
+  let word = randomWord();
+  let dissect = word.split('');
+
+  let regionalArray = [];
+
+  dissect.map(character => {
+    let regionalCharacter = ':regional_indicator_' + character + ':';
+    regionalArray.push(regionalCharacter);
+  })
+
+  let output = regionalArray.join('');
+  return evt.message.channel.sendMessage(output);
+}
+
+export default {
+  word
+};
+
+export const help = {
+  word
+};